Move SubmissionValues to crops
It's only being used there, makes sense to have it only there
