+++ frontend/src/Interface/CropsDetailsInterface.tsx
@@ -18,24 +18,4 @@ interface CropsDetailsInterface {
   willSawdustBeApplied: boolean | undefined;
 }
 
-/**
- * @summary     Interface for the main data file
- * @description This interface will be for Submission, it works the same as the CropsDetailInterface
- *              but this is only for the submission of formik.
- * @author      @Kcaparas
- */
-
-interface SubmissionCropsInterface {
-  id: number;
-  cropId: string;
-  yield: number;
-  plantAgeYears: string;
-  numberOfPlantsPerAcre: number;
-  distanceBtwnPlants: string;
-  distanceBtwnRows: string;
-  willPlantsBePruned: boolean | undefined;
-  whereWillPruningsGo: string;
-  willSawdustBeApplied: boolean | undefined;
-}
-
-export type { CropsDetailsInterface, SubmissionCropsInterface };
+export default CropsDetailsInterface;
